refactor(blacklist): move logging logic inside `BlacklistManager`
The changes in this commit focus on the logging of user blacklisting. The main change is the refactoring of the `logBlacklist` function from the `ModLogs` utility to the `BlacklistManager` class. This allows for more centralized and consistent handling of blacklist logging.

The key changes are:

- Removed the `logBlacklist` function from the `ModLogs` utility.
- Updated the `user.ts` command to use the `BlacklistManager.log` method instead of the old `logBlacklist` function.
- Removed the unused `logBlacklist` import from the `blacklistHandler.ts` and `server.ts` files.

This change ensures that the logging of user blacklisting is managed through the `BlacklistManager`, improving the overall maintainability and consistency of the codebase.
