+++ src/managers/BlacklistManager.ts
@@ -1,6 +1,12 @@
+import Constants, { emojis } from '#main/config/Constants.js';
 import BaseInfractionManager from '#main/managers/InfractionManager/BaseInfractionManager.js';
+import ServerInfractionManager from '#main/managers/InfractionManager/ServerInfractionManager.js';
+import db from '#main/utils/Db.js';
+import { sendLog } from '#main/utils/HubLogger/Default.js';
+import { resolveEval } from '#main/utils/Utils.js';
 import { InfractionStatus, Prisma, ServerInfraction, UserInfraction } from '@prisma/client';
-import { Snowflake } from 'discord.js';
+import { stripIndents } from 'common-tags';
+import { type Client, EmbedBuilder, type Snowflake, type User } from 'discord.js';
 
 export default class BlacklistManager<T extends UserInfraction | ServerInfraction> {
   readonly targetId: Snowflake;
@@ -74,5 +80,74 @@ export default class BlacklistManager<T extends UserInfraction | ServerInfractio
   public async fetchBlacklist(hubId: string) {
     const blacklist = await this.infracManager.fetchInfraction('BLACKLIST', hubId, 'ACTIVE');
     return blacklist;
+  }
+  /**
+   * Logs the blacklisting of a user or server.
+   * @param userOrServer - The user or server being blacklisted.
+   * @param mod - The moderator performing the blacklisting.
+   * @param reason - The reason for the blacklisting.
+   * @param expires - The optional expiration date for the blacklisting.
+   */
+  async log(
+    hubId: string,
+    client: Client,
+    opts: { mod: User; reason: string; expiresAt: Date | null },
+  ) {
+    const { mod, reason, expiresAt } = opts;
+
+    const hub = await db.hub.findFirst({ where: { id: hubId }, include: { logConfig: true } });
+    if (!hub?.logConfig[0].modLogs) return;
+
+    let name;
+    let iconURL;
+    let type;
+    let target;
+
+    if (this.infracManager instanceof ServerInfractionManager) {
+      target = resolveEval(
+        await client.cluster.broadcastEval(
+          (c, guildId) => {
+            const guild = c.guilds.cache.get(guildId);
+            if (!guild) return null;
+
+            return { name: guild.name, iconURL: guild.iconURL() ?? undefined, id: guildId };
+          },
+          { context: this.targetId },
+        ),
+      );
+      if (!target) return;
+
+      name = target.name;
+      iconURL = target.iconURL;
+      type = 'Server';
+    }
+    else {
+      target = await client.users.fetch(this.targetId);
+      name = target.username;
+      iconURL = target.displayAvatarURL();
+      type = 'User';
+    }
+
+    const embed = new EmbedBuilder()
+      .setAuthor({ name: `${type} ${name} blacklisted`, iconURL })
+      .setDescription(
+        stripIndents`
+				${emojis.dotBlue} **${type}:** ${name} (${target.id})
+				${emojis.dotBlue} **Moderator:** ${mod.username} (${mod.id})
+				${emojis.dotBlue} **Hub:** ${hub?.name}
+			`,
+      )
+      .addFields(
+        { name: 'Reason', value: reason, inline: true },
+        {
+          name: 'Expires',
+          value: expiresAt ? `<t:${Math.round(expiresAt.getTime() / 1000)}:R>` : 'Never.',
+          inline: true,
+        },
+      )
+      .setColor(Constants.Colors.interchatBlue)
+      .setFooter({ text: `Blacklisted by: ${mod.username}`, iconURL: mod.displayAvatarURL() });
+
+    await sendLog(opts.mod.client.cluster, hub?.logConfig[0].modLogs, embed);
   }
 }
