+++ src/utils/HubLogger/ModLogs.ts
@@ -6,83 +6,7 @@ import type { Hub, HubLogConfig } from '@prisma/client';
 import { stripIndents } from 'common-tags';
 import { type Client, codeBlock, EmbedBuilder, type Snowflake, User } from 'discord.js';
 import Constants, { emojis } from '../../config/Constants.js';
-import { resolveEval } from '../Utils.js';
 import { sendLog } from './Default.js';
-
-/**
- * Logs the blacklisting of a user or server.
- * @param userOrServer - The user or server being blacklisted.
- * @param mod - The moderator performing the blacklisting.
- * @param reason - The reason for the blacklisting.
- * @param expires - The optional expiration date for the blacklisting.
- */
-export const logBlacklist = async (
-  hubId: string,
-  client: Client,
-  opts: {
-    target: User | Snowflake;
-    mod: User;
-    reason: string;
-    expiresAt: Date | null;
-  },
-) => {
-  const { target: _target, mod, reason, expiresAt } = opts;
-
-  const hub = await db.hub.findFirst({ where: { id: hubId }, include: { logConfig: true } });
-  if (!hub?.logConfig[0].modLogs) return;
-
-  let name;
-  let iconURL;
-  let type;
-  let target;
-
-  if (typeof _target === 'string') {
-    target = resolveEval(
-      await client.cluster.broadcastEval(
-        (c, guildId) => {
-          const guild = c.guilds.cache.get(guildId);
-          if (!guild) return null;
-
-          return { name: guild.name, iconURL: guild.iconURL() ?? undefined, id: guildId };
-        },
-        { context: _target },
-      ),
-    );
-    if (!target) return;
-
-    name = target.name;
-    iconURL = target.iconURL;
-    type = 'Server';
-  }
-  else {
-    target = _target;
-    name = target.username;
-    iconURL = target.displayAvatarURL();
-    type = 'User';
-  }
-
-  const embed = new EmbedBuilder()
-    .setAuthor({ name: `${type} ${name} blacklisted`, iconURL })
-    .setDescription(
-      stripIndents`
-				${emojis.dotBlue} **${type}:** ${name} (${target.id})
-				${emojis.dotBlue} **Moderator:** ${mod.username} (${mod.id})
-				${emojis.dotBlue} **Hub:** ${hub?.name}
-			`,
-    )
-    .addFields(
-      { name: 'Reason', value: reason, inline: true },
-      {
-        name: 'Expires',
-        value: expiresAt ? `<t:${Math.round(expiresAt.getTime() / 1000)}:R>` : 'Never.',
-        inline: true,
-      },
-    )
-    .setColor(Constants.Colors.interchatBlue)
-    .setFooter({ text: `Blacklisted by: ${mod.username}`, iconURL: mod.displayAvatarURL() });
-
-  await sendLog(opts.mod.client.cluster, hub?.logConfig[0].modLogs, embed);
-};
 
 const getUnblacklistEmbed = (
   type: 'User' | 'Server',
